+++ client/src/routes/home/index.js
@@ -6,7 +6,12 @@ export default class Home extends Component {
 	render() {
 		return (
 			<div class={style.home}>
-				<SearchAutocomplete />
+				<p>What Can Map-e find for you today!</p>
+				<SearchAutocomplete /><button type="submit">Search</button>
+				<a href="/maps"><button>Map</button></a>
+				<a href="/directions"><button>Directions</button></a>
+				<a href="/pins"><button>Places</button></a>
+				<a href="/profile"><button>Profile</button></a>
 			</div>
 		);
 	}
