+++ client/src/routes/home/style.css
@@ -1,5 +1,8 @@
 .home {
 	padding: 56px 20px;
-	min-height: 100%;
 	width: 100%;
+}
+button {
+	display: block;
+	width: 175px;
 }
