+++ client/src/routes/profile/index.js
@@ -5,8 +5,7 @@ import SearchAutocomplete from '../../components/search';
 
 export default class Profile extends Component {
 	state = {
-		time: Date.now(),
-		count: 10
+		time: Date.now()
 	};
 
 	// gets called when this route is navigated to
@@ -25,20 +24,17 @@ export default class Profile extends Component {
 		this.setState({ time: Date.now() });
 	};
 
-
 	// Note: `user` comes from the URL, courtesy of our router
 	render({ user }, { time }) {
 		return (
 			<div class={style.profile}>
 				<h1>Profile: {user}</h1>
 				<p>This is the user profile for a user named { user }.</p>
-
 				<div>Current time: {new Date(time).toLocaleString()}</div>
 				<SearchAutocomplete style={{width: '90%'}}
-    onPlaceSelected={(place) => {
-      console.log(place);
-    }} 
-	
+    												onPlaceSelected={(place) => {
+      											console.log(place);
+    		}}
 	/>
 				<Favorites/>
 			</div>
